+++ CHANGELOG.md
@@ -35,6 +35,9 @@
 - Fixed bug where the `user` template alias was not set for the User Activated trigger. ([#31](https://github.com/doublesecretagency/craft-notifier/issues/31)) (thanks @chasegiunta)
 - Fixed `{% skipMessage %}` not being recognized inside the Twig sandbox. ([#29](https://github.com/doublesecretagency/craft-notifier/issues/29))
 
+### Removed
+- Removed the deprecated `twigSandbox` config setting. Use [`config/notifier-sandbox.php`](https://plugins.doublesecretagency.com/notifier/messages/twig-sandbox) instead.
+
 ## 2.1.0 - 2025-02-25
 
 ### Added